crm-5: using cache to store/retrieve user permission
